docs: add documentation for Security S2 (#3172)
