+++ packages/zwave-js/src/Controller.ts
@@ -4,5 +4,6 @@ export type {
 	ZWaveController,
 } from "./lib/controller/Controller";
 export type { ControllerStatistics } from "./lib/controller/ControllerStatistics";
+export * from "./lib/controller/Inclusion";
 export type { ZWaveLibraryTypes } from "./lib/controller/ZWaveLibraryTypes";
 export { RFRegion } from "./lib/serialapi/misc/SerialAPISetupMessages";
